test(kitchenpos): 주문 테스트 분리
