+++ src/test/java/kitchenpos/application/DeliveryOrderServiceTest.java
@@ -1,0 +1,93 @@
+package kitchenpos.application;
+
+import kitchenpos.domain.*;
+import kitchenpos.infra.KitchenridersClient;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
+
+import static kitchenpos.fixture.MenuFixture.menu;
+import static kitchenpos.fixture.OrderFixture.deliveryOrder;
+import static kitchenpos.fixture.OrderFixture.orderLineItem;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.mockito.AdditionalAnswers.returnsFirstArg;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
+import static org.mockito.Mockito.when;
+
+@SpringBootTest
+@ExtendWith(MockitoExtension.class)
+class DeliveryOrderServiceTest {
+
+    @MockBean
+    @Autowired
+    OrderRepository orderRepository;
+
+    @MockBean
+    @Autowired
+    MenuRepository menuRepository;
+
+    @MockBean
+    @Autowired
+    OrderTableRepository orderTableRepository;
+
+    @MockBean
+    @Autowired
+    KitchenridersClient kitchenridersClient;
+
+    @Autowired
+    OrderService orderService;
+
+    @DisplayName("배달 주문")
+    @Nested
+    class DeliveryOrder {
+        private Menu menu;
+        private long quantity;
+        private BigDecimal price;
+        private OrderLineItem orderLineItem;
+
+        @BeforeEach
+        void setUp() {
+            this.menu = menu();
+            this.quantity = 1L;
+            this.price = menu.getPrice().multiply(BigDecimal.valueOf(quantity));
+            this.orderLineItem = orderLineItem(menu, quantity, price);
+        }
+
+        @DisplayName("배달 주문을 생성할 수 있습니다.")
+        @Test
+        void createDeliveryOrder() {
+            final Menu menu = menu();
+            final long quantity = 1L;
+            final OrderLineItem orderLineItem = orderLineItem(menu, quantity, menu.getPrice().multiply(BigDecimal.valueOf(quantity)));
+            final String deliverAddress = "서울시 강남구";
+
+            when(menuRepository.findAllByIdIn(anyList())).thenReturn(List.of(menu));
+            when(menuRepository.findById(menu.getId())).thenReturn(Optional.ofNullable(menu));
+            when(orderRepository.save(any(Order.class))).then(returnsFirstArg());
+            final Order order = orderService.create(
+                    deliveryOrder(null, null, deliverAddress, OrderStatus.WAITING, List.of(orderLineItem))
+            );
+            assertAll(
+                    () -> assertThat(order.getId()).isNotNull(),
+                    () -> assertThat(order.getOrderDateTime()).isBeforeOrEqualTo(LocalDateTime.now()),
+                    () -> assertThat(order.getDeliveryAddress()).isEqualTo(deliverAddress),
+                    () -> assertThat(order.getStatus()).isEqualTo(OrderStatus.WAITING),
+                    () -> assertThat(order.getType()).isEqualTo(OrderType.DELIVERY),
+                    () -> assertThat(order.getOrderTable()).isNull()
+            );
+        }
+    }
+}
